Fix caching when using the s3 provider
The current configuration of the binary store for the s3 provider
causes Artifactory to download artifacts from s3 on every request. This
negatively impacts performance. This commit updates the binary store
configuration so that Artifactory first uses the cache-fs provider
before reaching out to s3. This change better aligns the configuration
of the binary store with the default configuration in the documentation
at https://www.jfrog.com/confluence/display/RTF/Configuring+the+Filestore#ConfiguringtheFilestore-S3ClusterBinaryProvider

This doesn't apply to the non-ha chart since it doesn't include a
binary store file